+++ libs/core/src/lib/message-strip/index.ts
@@ -1,1 +1,5 @@
-export * from './public_api';
+export * from './message-strip.module';
+export * from './message-strip.component';
+export * from './message-strip-type';
+export * from './auto-dismiss-message-strip.directive';
+export * from './alert';